Add ability to show containers for specific device
Moved ContainerMetas component to the 'Device' page directory. ContainerMetas has been given functionality to list containers specific to a device based on a provided device UUID. Likewise, routes are updated to link to this new device-specific container view. Other code cleanups and component restructures are also performed to improve maintainability.
