+++ client/src/pages/Devices/index.tsx
@@ -1,45 +1,22 @@
-import { DeviceStatType } from '@/components/Charts/DeviceStatType';
-import TinyLineDeviceGraph from '@/components/Charts/TinyLineDeviceGraph';
-import TinyRingProgressDeviceGraph from '@/components/Charts/TinyRingProgressDeviceGraph';
-import TinyRingProgressDeviceIndicator from '@/components/Charts/TinyRingProgressDeviceIndicator';
 import DeviceQuickActionDropDown from '@/components/DeviceComponents/DeviceQuickAction/DeviceQuickActionDropDown';
-import DeviceStatusTag from '@/components/DeviceComponents/DeviceStatusTag';
 import { OsLogo } from '@/components/DeviceComponents/OsLogo/OsLogo';
-import { CarbonBatchJob, WhhCpu, WhhRam } from '@/components/Icons/CustomIcons';
+import { CarbonBatchJob } from '@/components/Icons/CustomIcons';
 import Title, { PageContainerTitleColors } from '@/components/Template/Title';
 import TerminalModal, { TerminalStateProps } from '@/components/TerminalModal';
+import ListContent from '@/pages/Devices/components/ListComponent';
 import { getDevices } from '@/services/rest/device';
-import DeviceStatus from '@/utils/devicestatus';
-import { Link } from '@@/exports';
+import { Link } from '@umijs/max';
 import {
   AppstoreOutlined,
   ControlOutlined,
   TableOutlined,
 } from '@ant-design/icons';
 import { PageContainer } from '@ant-design/pro-components';
-import {
-  Avatar,
-  Button,
-  Card,
-  Carousel,
-  Col,
-  List,
-  Row,
-  Tooltip,
-  Typography,
-} from 'antd';
+import { Avatar, Button, Card, List, Tooltip } from 'antd';
 import React, { memo, useEffect, useState } from 'react';
 import { TerminalContextProvider } from 'react-terminal';
 import { API } from 'ssm-shared-lib';
 import styles from './Devices.less';
-
-const { Text } = Typography;
-
-export type StateType = {
-  visible?: boolean;
-  done?: boolean;
-  current?: API.DeviceItem;
-};
 
 const Index = memo(() => {
   const [deviceList, setDeviceList] = React.useState<API.DeviceList>({});
@@ -69,14 +46,6 @@ const Index = memo(() => {
   useEffect(() => {
     fetchDeviceList();
   }, []);
-  /*
-  const showModal = () => {
-    setState({
-      visible: true,
-      current: undefined,
-    });
-  };
-*/
 
   const onDropDownClicked = (key: string) => {
     switch (
@@ -85,116 +54,6 @@ const Index = memo(() => {
     ) {
     }
   };
-
-  const ListContent: React.FC<API.DeviceItem> = (props: API.DeviceItem) => (
-    <div className={styles.listContent} key={props.uuid}>
-      <div className={styles.listContentItem}>
-        <span>
-          <DeviceStatusTag status={props.status} />
-        </span>
-        <p>{props.hostname}</p>
-      </div>
-      <div className={styles.listContentItem} style={{ width: '80px' }}>
-        {props.status !== DeviceStatus.UNMANAGED && (
-          <>
-            <p style={{ minWidth: '80px' }}>
-              <WhhCpu /> {props.cpuSpeed?.toFixed(1)} Ghz
-            </p>
-            <p style={{ minWidth: '80px' }}>
-              <WhhRam /> {props.mem ? (props.mem / 1024).toFixed(0) : 'NaN'} Gb
-            </p>
-          </>
-        )}
-      </div>
-      <div className={styles.listContentItem}>
-        {(props.status !== DeviceStatus.UNMANAGED && (
-          <Carousel
-            style={{ width: 300, height: 70, zIndex: 1000 }}
-            dotPosition={'right'}
-            lazyLoad={'progressive'}
-          >
-            <div style={{ width: 300, height: 70, zIndex: 1000 }}>
-              <div
-                style={{
-                  width: 300,
-                  height: 70,
-                  zIndex: 1000,
-                  paddingTop: '15px',
-                }}
-              >
-                <Row style={{ alignItems: 'center' }} justify="center">
-                  <Col span={6}>
-                    <TinyRingProgressDeviceGraph
-                      type={DeviceStatType.CPU}
-                      deviceUuid={props.uuid}
-                    />
-                  </Col>
-                  <Col span={6}>
-                    <TinyRingProgressDeviceGraph
-                      type={DeviceStatType.MEM_USED}
-                      deviceUuid={props.uuid}
-                    />
-                  </Col>
-                  <Col span={6}>
-                    <TinyRingProgressDeviceIndicator
-                      deviceUuid={props.uuid}
-                      type={DeviceStatType.SERVICES}
-                    />
-                  </Col>
-                </Row>
-              </div>
-            </div>
-            <div style={{ width: 300, height: 70, zIndex: 1000 }}>
-              <div style={{ width: 300, height: 70, zIndex: 1000 }}>
-                <Row style={{ alignItems: 'center' }} justify="center">
-                  <Col span={24}>
-                    <TinyLineDeviceGraph
-                      type={DeviceStatType.CPU}
-                      deviceUuid={props.uuid}
-                      from={24}
-                    />
-                  </Col>
-                </Row>
-                <Row>
-                  <Col
-                    span={24}
-                    style={{ marginTop: '-5px', textAlign: 'center' }}
-                  >
-                    <Text type="secondary" style={{ fontSize: '8px' }}>
-                      CPU
-                    </Text>
-                  </Col>
-                </Row>
-              </div>
-            </div>
-            <div style={{ width: 300, height: 70, zIndex: 1000 }}>
-              <div style={{ width: 300, height: 70, zIndex: 1000 }}>
-                <Row style={{ alignItems: 'center' }} justify="center">
-                  <Col span={24}>
-                    <TinyLineDeviceGraph
-                      type={DeviceStatType.MEM_USED}
-                      deviceUuid={props.uuid}
-                      from={24}
-                    />
-                  </Col>
-                </Row>
-                <Row>
-                  <Col
-                    span={24}
-                    style={{ marginTop: '-5px', textAlign: 'center' }}
-                  >
-                    <Text type="secondary" style={{ fontSize: '8px' }}>
-                      MEM USED
-                    </Text>
-                  </Col>
-                </Row>{' '}
-              </div>
-            </div>
-          </Carousel>
-        )) || <div style={{ width: 300, height: 70, zIndex: 1000 }} />}
-      </div>
-    </div>
-  );
 
   return (
     <TerminalContextProvider>
@@ -239,16 +98,17 @@ const Index = memo(() => {
               renderItem={(item) => (
                 <List.Item
                   actions={[
-                    <a
-                      key={`showEditModal-${item.uuid}`}
-                      onClick={(e) => {
-                        e.preventDefault();
+                    <Link
+                      to={{
+                        pathname: `/manage/services`,
+                        search: `deviceUuid=${item.uuid}`,
                       }}
+                      key={`services-${item.uuid}`}
                     >
                       <Tooltip title="Services">
                         <AppstoreOutlined />
                       </Tooltip>
-                    </a>,
+                    </Link>,
                     <Link
                       to={`/admin/inventory/${item.uuid}`}
                       key={`devicesettings-${item.uuid}`}
@@ -270,7 +130,7 @@ const Index = memo(() => {
                     avatar={
                       <Avatar src={OsLogo(item.osLogoFile)} size="large" />
                     }
-                    title={<a href={item.hostname}>{item.hostname}</a>}
+                    title={item.hostname}
                     description={item.ip}
                   />
                   <ListContent
